fix(server): restore operator-snapshot failure visibility
Apply the root filter in SQL before LIMIT and keep non-propagating
PR-feedback child failures visible. Move snapshot tests off the
shrinking src/ inventory so CI placement can pass.
