+++ crates/harness-workflow/src/runtime/store/terminal_instance_queries.rs
@@ -62,6 +62,30 @@ impl WorkflowRuntimeStore {
         terminal_state: WorkflowTerminalState,
         limit: i64,
     ) -> anyhow::Result<Vec<WorkflowInstance>> {
+        self.list_recent_terminal_instances_filtered(definition_id, terminal_state, limit, false)
+            .await
+    }
+
+    /// Like [`Self::list_recent_terminal_instances_by_definition`], but only
+    /// root workflows. The `parent_workflow_id IS NULL` predicate runs before
+    /// `LIMIT` so newer child rows cannot crowd out older roots.
+    pub async fn list_recent_root_terminal_instances_by_definition(
+        &self,
+        definition_id: &str,
+        terminal_state: WorkflowTerminalState,
+        limit: i64,
+    ) -> anyhow::Result<Vec<WorkflowInstance>> {
+        self.list_recent_terminal_instances_filtered(definition_id, terminal_state, limit, true)
+            .await
+    }
+
+    async fn list_recent_terminal_instances_filtered(
+        &self,
+        definition_id: &str,
+        terminal_state: WorkflowTerminalState,
+        limit: i64,
+        roots_only: bool,
+    ) -> anyhow::Result<Vec<WorkflowInstance>> {
         let limit = limit.clamp(1, 500);
         let selectors = self
             .definition_registry
@@ -76,6 +100,7 @@ impl WorkflowRuntimeStore {
         let rows: Vec<(String, DateTime<Utc>)> = sqlx::query_as(
             "SELECT data::text, updated_at FROM workflow_instances
              WHERE definition_id = $1
+               AND (NOT $7 OR parent_workflow_id IS NULL)
                AND (
                    state = ANY($2::text[])
                    OR EXISTS (
@@ -96,6 +121,7 @@ impl WorkflowRuntimeStore {
         .bind(&query.definition_hashes)
         .bind(&query.versioned_states)
         .bind(limit)
+        .bind(roots_only)
         .fetch_all(&self.pool)
         .await?;
         rows.into_iter()
@@ -243,4 +269,74 @@ fn terminal_selector_query_parts(
         }
     }
     Ok(query)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::WorkflowRuntimeStore;
+    use crate::runtime::{
+        WorkflowInstance, WorkflowSubject, WorkflowTerminalState, GITHUB_ISSUE_PR_DEFINITION_ID,
+    };
+    use chrono::{Duration, Utc};
+    use harness_core::db::resolve_database_url;
+
+    #[tokio::test]
+    async fn root_terminal_listing_applies_parent_filter_before_limit() -> anyhow::Result<()> {
+        if resolve_database_url(None).is_err() {
+            return Ok(());
+        }
+
+        let dir = tempfile::tempdir()?;
+        let store = WorkflowRuntimeStore::open(&dir.path().join("workflow_runtime.db")).await?;
+        let root = WorkflowInstance::new(
+            GITHUB_ISSUE_PR_DEFINITION_ID,
+            1,
+            "failed",
+            WorkflowSubject::new("issue", "issue:root-limit"),
+        )
+        .with_id("root-failed-before-limit");
+        store.force_upsert_lifecycle_state_for_test(&root).await?;
+        sqlx::query("UPDATE workflow_instances SET updated_at = $2 WHERE id = $1")
+            .bind(&root.id)
+            .bind(Utc::now() - Duration::hours(1))
+            .execute(store.pool())
+            .await?;
+
+        for index in 0..20 {
+            let child = WorkflowInstance::new(
+                GITHUB_ISSUE_PR_DEFINITION_ID,
+                1,
+                "failed",
+                WorkflowSubject::new("issue", format!("issue:child-limit-{index}")),
+            )
+            .with_id(format!("child-failed-before-limit-{index}"))
+            .with_parent(&root.id);
+            store.force_upsert_lifecycle_state_for_test(&child).await?;
+        }
+
+        let unfiltered = store
+            .list_recent_terminal_instances_by_definition(
+                GITHUB_ISSUE_PR_DEFINITION_ID,
+                WorkflowTerminalState::Failed,
+                20,
+            )
+            .await?;
+        assert!(
+            unfiltered.iter().all(|workflow| workflow.id != root.id),
+            "unfiltered limit should be filled by newer children: {unfiltered:?}"
+        );
+
+        let roots = store
+            .list_recent_root_terminal_instances_by_definition(
+                GITHUB_ISSUE_PR_DEFINITION_ID,
+                WorkflowTerminalState::Failed,
+                20,
+            )
+            .await?;
+        assert!(
+            roots.iter().any(|workflow| workflow.id == root.id),
+            "root filter must run before limit so the older root remains: {roots:?}"
+        );
+        Ok(())
+    }
 }
